% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaCCRFunctions.R
\name{CNVPlotSeriesMeans}
\alias{CNVPlotSeriesMeans}
\title{Plot subjects' copy number data with cluster means overlayed for a single chromosome}
\usage{
CNVPlotSeriesMeans(Location, X, Cluster, NSubj = 3, lowery = -1,
  uppery = 1)
}
\arguments{
\item{Location}{A vector of length p with chromosomal locations}

\item{X}{A variable (p) by subject (n) data matrix}

\item{Cluster}{A vector of length p with cluster labels}

\item{NSubj}{A positve integer; number of randomly selected subjects to plot.}

\item{lowery, uppery}{Scalars; limits for y-axis}
}
\description{
Plot subjects' copy number data with cluster means overlayed for a single chromosome
}
\examples{
NULL
}

