% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaCCRFunctions.R
\name{GetClusters}
\alias{GetClusters}
\title{Compute Clusters from fusions}
\usage{
GetClusters(V)
}
\arguments{
\item{V}{An n by p-1 data matrix}
}
\description{
Compute Clusters from fusions
}
\examples{
NULL
}

