% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaCCRFunctions.R
\name{GetGammaCV}
\alias{GetGammaCV}
\title{Get optimal cross validated gamma values by various rules}
\usage{
GetGammaCV(ErrMat, rule = 1, gamma.seq)
}
\arguments{
\item{ErrMat}{Matrix of cross validated errors outputted by \code{GetCVErrMat}}

\item{rule}{A number indicating how optimal gamma should be chosen. 1 for minimum cv error,
2 for 1 standard error rule}

\item{gamma.seq}{sequence of regularization parameters used for cross validation.}
}
\value{
A scalar. Optimal gamma selected by CV rule.
}
\description{
Get optimal cross validated gamma values by various rules
}
\examples{
NULL
}

