% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaCCRFunctions.R
\name{GetParBlocks}
\alias{GetParBlocks}
\title{Function to compute blocks for parallization; should not be called directly}
\usage{
GetParBlocks(X, w)
}
\arguments{
\item{X}{An n by p data matrix}

\item{w}{A vector of positive scalars of length p-1}
}
\description{
Function to compute blocks for parallization; should not be called directly
}
\examples{
NULL
}

