% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaCCRFunctions.R
\name{ThreshV}
\alias{ThreshV}
\title{Threshold differences}
\usage{
ThreshV(V, X, mult = 1, thresh.value = NULL)
}
\arguments{
\item{V}{an n x p-1 matrix of differences}

\item{X}{an n x p matrix}

\item{mult}{scalar to multiply standard deviation}

\item{thresh.value}{optional user specified threshold value.}
}
\value{
VThreshed an n x p-1 matrix of thresholded differences
}
\description{
Threshold differences
}
\examples{
NULL
}

